Создает точку входа
